added users profiles section

+++ src/components/UploadChangeProPic.js
@@ -26,51 +26,60 @@ function UploadChangeProPic(props) {
         setUrlAvailable(true)
         setUploadBtn(true)
     }
+    const uploadUrl = ()=>{
+        setUploading(true)
+        const object = {
+            profileImg: {
+                userId: props.userId,
+                imageUrl: srcUrl
+            }
+        }
+        // adding img url to the database
+        firebase.database().ref(`/${props.userId}`).update(object)
+        //  once successfull, the following functions are called. 
+        setUploading(false)
+        setSrcUrl('')
+        setUrlAvailable(false)
+        // turningoff the modal
+        props.modalWindow()
+
+    }
     const handleUploadImg = async (e)=>{
-        const imageObj = props.user.profileImg;
         if(urlAvailable){
             // first check if user previously had uploaded a photo before. if yes then we have to delete the old file and then update the data base
-            if (imageObj.imageId){
-                // creating the object to add to the database
-                const object = {
-                    profileImg: {
-                        userId: props.userId,
-                        imageUrl: srcUrl
+            if (props.user.profileImg ){
+                const imageObj = props.user.profileImg;
+                if (imageObj.imageId){
+                    setUploading(true)
+
+                    // creating the object to add to the database
+                    const object = {
+                        profileImg: {
+                            userId: props.userId,
+                            imageUrl: srcUrl
+                        }
                     }
+                    const storageRef = firebase.storage().ref('image').child(imageObj.imageId);
+                    // deleting the old photo 
+                    await storageRef.delete().then(() => {
+                        // then delete data of the image from the data base which is the just the id 
+                        firebase.database().ref(`/${props.userId}`).child('profileImg').child('imageId').remove()
+                    });
+                    // adding img url to the database
+                    firebase.database().ref(`/${props.userId}`).update(object);
+                    //  once successfull, the following functions are called. 
+                    setUploading(false)
+                    setSrcUrl('')
+                    setUrlAvailable(false)
+                    // turningoff the modal
+                    props.modalWindow()
+                    
+                } else {
+                    uploadUrl()
                 }
-                const storageRef = firebase.storage().ref('image').child(imageObj.imageId);
-                // deleting the old photo 
-                await storageRef.delete().then(() => {
-                    // then delete data of the image from the data base which is the just the id 
-                    firebase.database().ref(`/${props.userId}`).child('profileImg').child('imageId').remove()
-                });
-                // adding img url to the database
-                firebase.database().ref(`/${props.userId}`).update(object);
-                //  once successfull, the following functions are called. 
-                setUploading(false)
-                setSrcUrl('')
-                setUrlAvailable(false)
-                // turningoff the modal
-                props.modalWindow()
                 
             } else {
-                setUploading(true)
-                const object = {
-                    profileImg: {
-                        userId: props.userId,
-                        imageUrl: srcUrl
-                    }
-                }
-                // adding img url to the database
-
-                firebase.database().ref(`/${props.userId}`).update(object)
-                //  once successfull, the following functions are called. 
-                setUploading(false)
-                setSrcUrl('')
-                setUrlAvailable(false)
-                // turningoff the modal
-                props.modalWindow()
-
+                uploadUrl()
             }
         }else {
             // providing each images a unique id
@@ -81,6 +90,8 @@ function UploadChangeProPic(props) {
             // const fileRef = storageRef.child(file.name);
             if (props.user.profileImg ){
                 // this part is when user has a profile photo but want to change it. 
+                const imageObj = props.user.profileImg;
+
                 // turn on loading icon 
                 setUploading(true)
                 // Check if the image has an image id, since the user might have just added an url instead of upload an image.
